# apps/api/docs/swagger.yaml
basePath: /
definitions:
...
  httpx.ErrorResponse:
    properties:
      code:
        example: not_found
        type: string
      error:
        example: visit not found
        type: string
    type: object
...
  navigation.NavEdge:
    properties:
      accessible:
        type: boolean
      distance:
        type: number
      edgeType:
        description: CORRIDOR, ELEVATOR, STAIRS
        type: string
      fromNodeId:
        type: string
      id:
        type: string
      toNodeId:
        type: string
    type: object
  navigation.NavNode:
    properties:
      floorId:
        type: string
      id:
        type: string
      nodeType:
        description: ENTRANCE, PLACE_ENTRY, CORRIDOR, ELEVATOR, STAIRS
        type: string
      x:
        type: number
      "y":
        type: number
      zone:
        type: string
    type: object
  navigation.Route:
    properties:
      nodes:
        items:
          $ref: '#/definitions/navigation.NavNode'
        type: array
      segments:
        items:
          $ref: '#/definitions/navigation.NavEdge'
        type: array
      totalDistance:
        type: number
    type: object
...
info:
  contact: {}
  description: |-
    Patient journey and indoor navigation API above the HIS.
    Serves the CarePath-derived journey plan (ADR-0009) with
    every actionable step resolved to its service point.
  title: CarePath API
  version: 1.0.0
paths:
...
  /api/v1/navigation/edges:
    get:
      description: '#105/ADR-0015: the graph''s directed, costed edges. Two-way connections
        are two rows (A→B and B→A) and a cross-floor transition is an edge whose endpoints
        sit on different floors, so this is the whole walkable graph, not a per-floor
        slice.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/navigation.NavEdge'
            type: array
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: List navigation edges
      tags:
      - navigation
  /api/v1/navigation/nodes:
    get:
      description: '#105/ADR-0015: the graph''s nodes, which apps/web used to read
        from a build-time copy of packages/floorplans/graphs. The QR stickers staff
        print for lifts, stairs and entrances are derived from this list, so a stale
        copy sends patients to nodes that no longer exist — which is why the copy
        had to go.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/navigation.NavNode'
            type: array
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: List navigation nodes
      tags:
      - navigation
  /api/v1/navigation/route:
    get:
      description: Shortest walkable route for the SVG overlay. from is the current
        location's navigation node id — the nodeId of a location observation (e.g.
        I-1301/node-reception); to is the destination service point's code, which
        the server resolves to the place's entry node before routing.
      parameters:
      - description: Current location node id (e.g. I-1301/node-reception)
        in: query
        name: from
        required: true
        type: string
      - description: Destination service point code (e.g. PHARMACY)
        in: query
        name: to
        required: true
        type: string
      - description: Skip non-accessible edges (stairs) so the route detours via the
          elevator (FR-20)
        in: query
        name: accessibleOnly
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/navigation.Route'
        "400":
          description: missing from or to, or accessibleOnly is not a boolean
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: unknown node or service point, destination without a mapped
            place, or no walkable path
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Get the walking route to a service point
      tags:
      - navigation
...
swagger: "2.0"